+++ lib/chewy/railtie.rb
@@ -1,22 +1,53 @@
 module Chewy
   class Railtie < Rails::Railtie
-    NOTIFICATOR = ->(action, name, start, finish, id, payload) do
-      duration = ((finish - start).to_f * 10000).round / 10.0
-      Rails.logger.debug("  \e[1m\e[32m#{payload[:type].presence || payload[:index]} #{action} (#{duration}ms)\e[0m #{payload[:import]}")
+    module ControllerPatch
+      extend ActiveSupport::Concern
+      included do
+        prepend_around_action :setup_chewy_strategy
+      end
+
+    private
+
+      def setup_chewy_strategy
+        Chewy.strategy(:atomic) { yield }
+      end
     end
 
     rake_tasks do
       load 'tasks/chewy.rake'
     end
 
-    initializer 'chewy.add_app_chewy_path' do |app|
-      app.config.paths.add 'app/chewy'
+    console do |app|
+      if app.sandbox?
+        Chewy.strategy(:bypass)
+      else
+        Chewy.strategy(:urgent)
+      end
     end
 
-    initializer 'chewy.add_requests_logging' do |app|
-      ActiveSupport::Notifications.subscribe('import_objects.chewy', &NOTIFICATOR.curry['Import'])
-      ActiveSupport::Notifications.subscribe('search_query.chewy', &NOTIFICATOR.curry['Search'])
-      ActiveSupport::Notifications.subscribe('delete_query.chewy', &NOTIFICATOR.curry['Delete by Query'])
+    initializer 'chewy.migrations_patch' do
+      ActiveSupport.on_load(:active_record) do
+        ActiveRecord::Migration.class_eval do
+          def migrate_with_chewy(*_)
+            Chewy.strategy(:bypass) { migrate_without_chewy(*_) }
+          end
+          alias_method_chain :migrate, :chewy
+        end
+      end
+    end
+
+    initializer 'chewy.action_wrapper' do
+      ActiveSupport.on_load(:action_controller) do
+        include ControllerPatch
+      end
+    end
+
+    initializer 'chewy.logger' do |app|
+      Chewy.logger ||= ::Rails.logger
+    end
+
+    initializer 'chewy.add_app_chewy_path' do |app|
+      app.config.paths.add 'app/chewy'
     end
   end
 end