+++ lib/chewy/strategy/base.rb
@@ -8,6 +8,9 @@ module Chewy
     #   end
     #
     class Base
+      def name
+        self.class.name.demodulize.underscore.to_sym
+      end
       # This method called when some model tries to update index
       #
       def update type, objects, options = {}